App/backend/requirements.txt
fastapi[all]
uvicorn
tensorflow
pillow
numpy
pydantic
ultralytics
